# api/openapi.yaml
openapi: 3.0.3
info:
  title: openbot API
  version: 0.1.0
  description: |
    HTTP API for openbotd. Authenticated with a session cookie (`openbot_session`).
    Mutating requests must send `X-Openbot-CSRF: 1` (checked together with the Origin header).

    Realtime updates are delivered over a WebSocket at `GET /api/ws`; each frame is a JSON
    `WsEvent` (see components). The WebSocket is not described as an operation because
    OpenAPI can't express it; its payloads are.
servers:
  - url: /api
paths:
...
  /auth/setup:
    post:
      operationId: setupAccount
      tags: [auth]
      security: []
      description: Create the single user account. Only allowed while no account exists.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "200":
          description: Account created and session started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
...
  /chats:
    get:
      operationId: listChats
      tags: [chats]
      responses:
        "200":
          description: All chats, most recently active first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Chat" }
        "401": { $ref: "#/components/responses/Error" }
  /sidebar/categories:
    get:
      operationId: listSidebarCategories
      tags: [chats]
      description: The user's sidebar categories, in order. Only the user arranges the sidebar.
      responses:
        "200":
          description: Categories in order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SidebarCategory" }
        "401": { $ref: "#/components/responses/Error" }
    post:
      operationId: createSidebarCategory
      tags: [chats]
      description: Add a category at the end. Broadcasts sidebar.updated.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSidebarCategoryRequest" }
      responses:
        "201":
          description: The new category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SidebarCategory" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /sidebar/categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateSidebarCategory
      tags: [chats]
      description: Rename or collapse/expand a category. Broadcasts sidebar.updated.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSidebarCategoryRequest" }
      responses:
        "200":
          description: The category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SidebarCategory" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteSidebarCategory
      tags: [chats]
      description: Delete a category; its chats move to Unassigned. Broadcasts sidebar.updated.
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /sidebar/layout:
    put:
      operationId: setSidebarLayout
      tags: [chats]
      description: |
        Arrange the sidebar after a drag: the order of categories, and for each listed section
        (a category, or null for Unassigned) its chats in order. Chats not listed keep their
        place. Broadcasts sidebar.updated.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SidebarLayout" }
      responses:
        "204": { description: Saved }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
...
  /connectors/catalog:
    get:
      operationId: listConnectorCatalog
      tags: [connectors]
      description: Apps that connect with a sign-in (their official MCP servers), ready to pick.
      responses:
        "200":
          description: The catalog
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CatalogApp" }
        "401": { $ref: "#/components/responses/Error" }
  /connectors/sign-in:
    post:
      operationId: startSignIn
      tags: [connectors]
      description: |
        Start signing in to an MCP server that uses OAuth: for a new connection (url, name,
        agentIds), a connect card in chat (messageId) or reconnecting a connection
        (connectorId). Send the user to authorizeUrl. They come back to /oauth/callback, which
        finishes and redirects into the app. When pasteBack is true the server wouldn't accept
        openbot's address: the user lands on an error page and pastes its address into
        /connectors/sign-in/complete.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StartSignInRequest" }
      responses:
        "200":
          description: Where to send the user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StartSignInResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: The connect card was already answered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: openbot couldn't reach the server
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /connectors/sign-in/complete:
    post:
      operationId: completeSignIn
      tags: [connectors]
      description: Finish a paste-back sign-in with the address of the page the user landed on.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompleteSignInRequest" }
      responses:
        "200":
          description: The connection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignInResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "502":
          description: openbot couldn't reach the server
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: openbot_session
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code:
          type: string
          description: |
            Machine-readable reason where clients react to it: sign_in_required (an MCP server
            wants a sign-in instead of a key), reconnect (a sign-in expired)
...
    CatalogApp:
      type: object
      required: [id, name, description, url]
      properties:
        id: { type: string, description: 'e.g. "linear"' }
        name: { type: string }
        description: { type: string }
        url: { type: string, description: The app's MCP server }
    StartSignInRequest:
      type: object
      properties:
        url: { type: string, description: The MCP server (new connections) }
        name: { type: string }
        agentIds:
          type: array
          items: { type: string }
        messageId: { type: string, description: A connect card to answer }
        connectorId: { type: string, description: A connection to reconnect }
    StartSignInResponse:
      type: object
      required: [authorizeUrl, pasteBack]
      properties:
        authorizeUrl: { type: string }
        pasteBack:
          type: boolean
          description: The user will land on an error page whose address they paste back
    CompleteSignInRequest:
      type: object
      required: [callbackUrl]
      properties:
        callbackUrl: { type: string, description: The address of the page the user landed on }
    SignInResult:
      type: object
      required: [connector, chatId]
      properties:
        connector: { $ref: "#/components/schemas/Connector" }
        chatId:
          type: string
          nullable: true
          description: The chat whose connect card this answered
    Credentials:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 12, maxLength: 256 }
    User:
      type: object
      required: [id, username]
      properties:
        id: { type: string }
        username: { type: string }
...
    Connector:
      type: object
      required: [id, type, name, config, credentialsSet, agentIds, webhookUrl, webhookSecret, signIn, createdAt]
      properties:
        id: { type: string }
        type: { type: string }
        name: { type: string }
        config:
          type: object
          additionalProperties: { type: string }
        credentialsSet:
          type: array
          description: Credential keys that have a saved value (values are never returned)
          items: { type: string }
        agentIds:
          type: array
          description: Agents allowed to use this account
          items: { type: string }
        webhookUrl:
          type: string
          nullable: true
          description: Where the service should send events (types with webhooks)
        webhookSecret:
          type: string
          nullable: true
          description: |
            A secret openbot generated for the service's webhook settings (e.g. GitHub's), shown so
            the user can copy it there
        signIn:
          type: string
          nullable: true
          enum: [ok, expired]
          description: Connected by signing in (MCP OAuth); expired means it needs Reconnect
        createdAt: { type: string, format: date-time }
...
    ChatMember:
      type: object
      required: [agentId, position]
      properties:
        agentId: { type: string }
        position: { type: integer }
    Chat:
      type: object
      required: [id, kind, name, members, unreadCount, lastMessage, createdAt, categoryId, position]
      properties:
        id: { type: string }
        kind: { type: string, enum: [dm, group] }
        name:
          type: string
          description: Group name, or the agent's name for DMs
        members:
          type: array
          description: Agent participants (the user is implicit in every chat)
          items: { $ref: "#/components/schemas/ChatMember" }
        unreadCount: { type: integer }
        lastMessage:
          allOf: [{ $ref: "#/components/schemas/Message" }]
          nullable: true
        createdAt: { type: string, format: date-time }
        categoryId:
          type: string
          nullable: true
          description: The sidebar category it's in; null = Unassigned
        position:
          type: integer
          nullable: true
          description: |
            Its place within its section once the user arranged it. Within a section, chats
            with no position (new or never moved) come first, most recently active first, then
            the rest by position.
    SidebarCategory:
      type: object
      required: [id, name, collapsed]
      properties:
        id: { type: string }
        name: { type: string }
        collapsed: { type: boolean }
    CreateSidebarCategoryRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
    UpdateSidebarCategoryRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
        collapsed: { type: boolean }
    SidebarLayout:
      type: object
      required: [categoryOrder, sections]
      properties:
        categoryOrder:
          type: array
          description: Every category id, in the new order
          items: { type: string }
        sections:
          type: array
          description: Sections whose chats changed, each with all its chats in order
          items:
            type: object
            required: [categoryId, chatIds]
            properties:
              categoryId:
                type: string
                nullable: true
                description: null for Unassigned
              chatIds:
                type: array
                items: { type: string }
    MessageAuthor:
      type: object
      required: [kind, agentId]
      properties:
        kind: { type: string, enum: [user, agent, system] }
        agentId:
          type: string
          nullable: true
          description: The agent that wrote it; null on an agent message means the agent was deleted
    Message:
      type: object
      required: [id, chatId, author, body, createdAt, reactions, mentions, attachments]
      properties:
        id:
          type: string
          description: ULID; lexicographic order is chronological order
        chatId: { type: string }
        author: { $ref: "#/components/schemas/MessageAuthor" }
        body:
          type: string
          description: Markdown
        createdAt: { type: string, format: date-time }
        failure:
          allOf: [{ $ref: "#/components/schemas/MessageFailure" }]
          nullable: true
          description: Set on system messages reporting that an agent's turn failed
        mentions:
          type: array
          description: Agents @mentioned in the message (group chats)
          items: { type: string }
        attachments:
          type: array
          description: Files and images attached to the message, in order
          items: { $ref: "#/components/schemas/Attachment" }
        reactions:
          type: array
          description: Emoji reactions, in the order each emoji was first used
          items: { $ref: "#/components/schemas/Reaction" }
        prompt:
          allOf: [{ $ref: "#/components/schemas/Prompt" }]
          nullable: true
          description: |
            Set on agent messages that ask the user something with clickable answers. `body`
            holds the question as plain text (for previews).
        event:
          allOf: [{ $ref: "#/components/schemas/MessageEvent" }]
          nullable: true
          description: Set on system messages that mark something that happened, e.g. an agent was created
        clientId:
          type: string
          description: |
            Echo of SendMessageRequest.clientId. Only present on the sendMessage response and the
            matching message.created event.
    MessageFailure:
      type: object
      required: [agentId, reason, retryable]
      properties:
        agentId: { type: string }
        reason:
          type: string
          enum: [no_key, invalid_key, model_blocked, provider_error, too_many_steps]
          description: |
            model_blocked: OpenCode refuses the agent's model (its provider trains on request
            data and the workspace's privacy settings forbid that). Offer to change the model.
        retryable: { type: boolean }
...
    Reaction:
      type: object
      required: [emoji, by]
      properties:
        emoji: { type: string }
        by:
          type: array
          description: Who reacted with this emoji, oldest first
          items: { $ref: "#/components/schemas/MessageAuthor" }
    Prompt:
      type: object
      required: [kind, question, options, allowOther, status, answer]
      properties:
        kind:
          type: string
          enum: [single, multi, text, approval, connect]
          description: |
            single = pick one; multi = pick any number; text = free-text answer; approval = the
            agent wants to do something that needs the user's OK (options: Approve, Decline, and
            sometimes a third, "Always allow", which approves and stops asking for that action);
            connect = the agent proposes connecting an app (see connection; options: Connect,
            Decline; connect with POST /messages/{messageId}/connect)
        question: { type: string }
        options:
          type: array
          items: { $ref: "#/components/schemas/PromptOption" }
        allowOther:
          type: boolean
          description: Offer "type your own" in addition to the options
        status: { type: string, enum: [pending, answered, dismissed] }
        answer:
          allOf: [{ $ref: "#/components/schemas/PromptAnswer" }]
          nullable: true
        connection:
          allOf: [{ $ref: "#/components/schemas/PromptConnection" }]
          nullable: true
          description: The proposed connection (connect prompts only)
        preview:
          allOf: [{ $ref: "#/components/schemas/ActionPreview" }]
          nullable: true
          description: What an approval prompt would do, for the card (approval prompts only)
...
    PromptOption:
      type: object
      required: [label]
      properties:
        label: { type: string }
        opensChatId:
          type: string
          description: Choosing this option also opens this chat (e.g. "Go talk to X")
    PromptAnswer:
      type: object
      properties:
        selected:
          type: array
          description: Indexes of the chosen options
          items: { type: integer, minimum: 0 }
        text:
          type: string
          maxLength: 4000
          description: Free-text answer (text prompts, or "type your own")
    MessageEvent:
      type: object
      required: [kind, agentId]
      properties:
        kind: { type: string, enum: [agent_created] }
        agentId:
          type: string
          description: The agent the event is about (e.g. the one that was created)
        chatId:
          type: string
          description: A chat to link to (e.g. the new agent's DM)
...
    Attachment:
      type: object
      required: [id, name, mime, size, url, width, height]
      properties:
        id: { type: string }
        name: { type: string, description: The file name }
        mime: { type: string, description: 'e.g. "image/png", "application/pdf"' }
        size: { type: integer, format: int64, description: Bytes }
        url: { type: string, description: Where to fetch it (same origin, needs the session) }
        width: { type: integer, nullable: true, description: Pixels, for images }
        height: { type: integer, nullable: true, description: Pixels, for images }
...
security:
  - session: []